requirements.txt
# requirements.txt

# Core Data Science & Time Series Libraries
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
matplotlib==3.9.0
seaborn==0.13.2
scikit-learn==1.5.0

# LangChain and Google Generative AI Libraries
langchain==0.2.5
langchain-google-genai==1.0.6
langchain-core==0.2.9

# Environment Variable Management
python-dotenv==1.0.1